[OF-1672] fix: Moved Callhelpers back to root (Core), and broke RealtimeEngine dependency by removing calls to INVOKE_IF_NOT_NULL
