+++ Library/src/Systems/Users/UserSystem.cpp
@@ -31,7 +31,7 @@
 #include "Systems/ResultHelpers.h"
 #include "Systems/Users/Authentication.h"
 
-#include <Common/CallHelpers.h>
+#include <CallHelpers.h>
 
 namespace chs_user = csp::services::generated::userservice;
 namespace chs_aggregation = csp::services::generated::aggregationservice;
